requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
tensorflow
numpy
pillow
python-multipart
fastapi-mail
passlib[bcrypt]
python-jose
